Add -Wconversion flag in makefile And switch to clang over gcc
- With Wconversion it's easier to change datatypes and don't forget
about them in whole code
- gcc is too strict about type conversions and clangd vs gcc gives
different warnings which are difficult to fix

+++ src/controls.c
@@ -37,7 +37,7 @@ void draw_button_cb(struct binded_widget_pair_t *bind) {
   draw_square(cr, colors, pixels, draw_ctx->palette);
 
   /* Actually redraw */
-  gtk_widget_queue_draw_area(drawing_area, 0, 0, pixels, pixels);
+  gtk_widget_queue_draw_area(drawing_area, 0, 0, (gint) pixels, (gint) pixels);
 }
 
 void compute_button_cb(struct computation_context_t *ctx) {
